+++ inc/deprecated.php
@@ -9,6 +9,76 @@
 if ( ! defined( 'ABSPATH' ) ) exit;
 
 /**
+ * Returns header page ID
+ *
+ * @since 1.1.1
+ */
+if ( ! function_exists( 'oceanwp_header_page_id' ) ) {
+
+	function oceanwp_header_page_id() {
+
+		// Return false if custom header is not selected
+		if ( 'custom' != oceanwp_header_style() ) {
+			return false;
+		}
+
+		// Get page ID from Customizer
+		$page_id = get_theme_mod( 'ocean_header_page_id' );
+
+		// Get page content
+		if ( ! empty( $page_id ) ) {
+
+			$page = get_post( $page_id );
+
+			if ( $page && ! is_wp_error( $page ) ) {
+				$page_id = $page->post_content;
+			}
+
+		}
+
+		// Apply filters and return content
+		return apply_filters( 'ocean_header_page_id', $page_id );
+
+	}
+
+}
+
+/**
+ * Returns footer page ID
+ *
+ * @since 1.0.0
+ */
+if ( ! function_exists( 'oceanwp_footer_page_id' ) ) {
+
+	function oceanwp_footer_page_id() {
+
+		// Return false if disabled via Customizer
+		if ( true != get_theme_mod( 'ocean_footer_widgets', true ) ) {
+			return null;
+		}
+
+		// Get page ID from Customizer
+		$page_id = get_theme_mod( 'ocean_footer_widgets_page_id' );
+
+		// Get page content
+		if ( ! empty( $page_id ) ) {
+
+			$page = get_post( $page_id );
+
+			if ( $page && ! is_wp_error( $page ) ) {
+				$page_id = $page->post_content;
+			}
+
+		}
+
+		// Apply filters and return content
+		return apply_filters( 'ocean_footer_page_id', $page_id );
+
+	}
+
+}
+
+/**
  * Custom excerpts based on wp_trim_words
  *
  * @since	1.0.0