+++ inc/helpers.php
@@ -778,41 +778,6 @@ if ( ! function_exists( 'oceanwp_header_classes' ) ) {
 }
 
 /**
- * Returns header page ID
- *
- * @since 1.1.1
- */
-if ( ! function_exists( 'oceanwp_header_page_id' ) ) {
-
-	function oceanwp_header_page_id() {
-
-		// Return false if custom header is not selected
-		if ( 'custom' != oceanwp_header_style() ) {
-			return false;
-		}
-
-		// Get page ID from Customizer
-		$page_id = get_theme_mod( 'ocean_header_page_id' );
-
-		// Get page content
-		if ( ! empty( $page_id ) ) {
-
-			$page = get_post( $page_id );
-
-			if ( $page && ! is_wp_error( $page ) ) {
-				$page_id = $page->post_content;
-			}
-
-		}
-
-		// Apply filters and return content
-		return apply_filters( 'ocean_header_page_id', $page_id );
-
-	}
-
-}
-
-/**
  * Returns retina header logo
  *
  * @since 1.1.1
@@ -3045,41 +3010,6 @@ if ( ! function_exists( 'oceanwp_footer_classes' ) ) {
 
 		// return classes
 		return $classes;
-
-	}
-
-}
-
-/**
- * Returns footer page ID
- *
- * @since 1.0.0
- */
-if ( ! function_exists( 'oceanwp_footer_page_id' ) ) {
-
-	function oceanwp_footer_page_id() {
-
-		// Return false if disabled via Customizer
-		if ( true != get_theme_mod( 'ocean_footer_widgets', true ) ) {
-			return null;
-		}
-
-		// Get page ID from Customizer
-		$page_id = get_theme_mod( 'ocean_footer_widgets_page_id' );
-
-		// Get page content
-		if ( ! empty( $page_id ) ) {
-
-			$page = get_post( $page_id );
-
-			if ( $page && ! is_wp_error( $page ) ) {
-				$page_id = $page->post_content;
-			}
-
-		}
-
-		// Apply filters and return content
-		return apply_filters( 'ocean_footer_page_id', $page_id );
 
 	}